LayoutTests/http/wpt/wk-web-animations/interfaces/element-get-animations.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Element.getAnimations()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

let a, b, c;

const targetA = document.createElement("a");
const targetB = document.createElement("b");

const effectAa = new KeyframeEffect(targetA, null);
const effectAb = new KeyframeEffect(targetA, null);
const effectB = new KeyframeEffect(targetB, null);

test(t => {
    assert_true(typeof targetA.getAnimations === "function");
}, "Element exposes the getAnimations method.");

test(t => {
    assert_array_equals(targetA.getAnimations(), []);
    assert_array_equals(targetB.getAnimations(), []);
}, "Element.getAnimations() returns an empty array when no animations were added.");

test(t => {
    a = new Animation(effectAa);
    b = new Animation(effectAb);
    c = new Animation(effectB);
    assert_array_equals(targetA.getAnimations(), [a, b]);
    assert_array_equals(targetB.getAnimations(), [c]);
}, "Element.getAnimations() returns an array with animations targetting that element.");

test(t => {
    b.effect = effectB;
    assert_array_equals(targetA.getAnimations(), [a]);
    assert_array_equals(targetB.getAnimations(), [b]);
}, "An effect can only be applied to a single animation and changing it changes the animations returned by Element.getAnimations().");

test(t => {
    a.effect = null;
    b.effect = null;
    c.effect = null;
    assert_array_equals(targetA.getAnimations(), []);
    assert_array_equals(targetB.getAnimations(), []);
}, "Setting an animation's effect to null changes the animations returned by Element.getAnimations().");

</script>
</body>
